C0001_Sql_Executor/requirements.txt
fastapi
pydantic
typing
uvicorn
psycopg2-binary